requirements.txt
# Core Dependencies
selenium==4.15.2
undetected-chromedriver==3.5.4
pyautogui==0.9.54
setuptools==69.0.3

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# PDF Processing for CV Matching
PyPDF2==3.0.1

# Data Processing
# numpy==1.24.3
# Numpy a fost eliminat din dependente pentru compatibilitate cu Python 3.13.
# Modulul de anti-detection foloseste acum un fallback care nu necesita numpy.

# JobMatcher (Optiunea A - fuzzy + TF-IDF + keywords)
scikit-learn>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# Environment Variables
python-dotenv>=1.0.0

# LLM / Fallback providers
openai>=1.0.0
google-generativeai>=0.8.0
mistralai>=1.0.0
requests>=2.28.0

# Standard Library (included with Python, listed for reference)
# csv
# json
# pathlib
# hashlib
# datetime
# random
# time
# os
# sys
# sqlite3
# threading
